requirements.txt
yt-dlp
python-telegram-bot[job-queue]
python-dotenv
requests
responses